+++ starboard/nplb/posix_compliance/posix_socket_join_multicast_group_test.cc
@@ -106,7 +106,7 @@ TEST(PosixSocketJoinMulticastGroupTest, SunnyDay) {
       FAIL() << "Failed to send multicast packet: " << errno;
       return;
     }
-    EXPECT_EQ(sizeof(kBuf), sent);
+    EXPECT_EQ(static_cast<ssize_t>(sizeof(kBuf)), sent);
     break;
   }
 
@@ -126,7 +126,7 @@ TEST(PosixSocketJoinMulticastGroupTest, SunnyDay) {
       usleep(1000);
       continue;
     }
-    EXPECT_EQ(sizeof(kBuf), received);
+    EXPECT_EQ(static_cast<ssize_t>(sizeof(kBuf)), received);
     break;
   }
